+++ lib/mutant/reporter/cli/printer/status_progressive.rb
@@ -14,7 +14,6 @@ module Mutant
             :amount_mutations_alive,
             :amount_mutations_killed,
             :killtime,
-            :overhead,
             :runtime
           )