+++ spec/unit/mutant/reporter/cli/printer/env_result_spec.rb
@@ -36,7 +36,7 @@ RSpec.describe Mutant::Reporter::CLI::Printer::EnvResult do
       Timeouts:        0
       Runtime:         4.00s
       Killtime:        2.00s
-      Overhead:        100.00%
+      Efficiency:      50.00%
       Mutations/s:     0.50
       Coverage:        50.00%
     STR